commit to version 1.2

+++ provinder_8081/src/main/java/com/grace/it/dao/DeptDao.java
@@ -23,5 +23,5 @@ public interface DeptDao {
 
     public Integer updateByCondition(Dept dept);
 
-    public Long insert(Dept dept);
+    public Integer insert(Dept dept);
 }
