+++ provinder_8081/src/main/java/com/grace/it/dao/DeptManagerDao.java
@@ -25,7 +25,7 @@ public interface DeptManagerDao {
 
     public Integer updateByCondition(DeptManager deptManager);
 
-    public Long insert(DeptManager deptManager);
+    public Integer insert(DeptManager deptManager);
 
     public Integer deleteByEmpNo(Long empNo);
 }
